refactor: move check group cmds into their own module
